:tada: Maybe have GHA freeze issues sorted?!
...  For _reasons_ the `gha-utilities/freeze-first-code-blocks` Action
was writing files owned by `root:root` with permissions of `600`, which
the `actions/upload-pages-artifact` Action took no liking to.

Forcing `gha-utilities/freeze-first-code-blocks` to add read permissions
via `chmod go+r <path>` seems to have resolved `tar Permission Denied`
sorts of errrors...  for now.  However, this may be a possible bug in
`freeze` or an issue with Docker runner as none of the other image
manipulation steps seem to suffer from this behavior :shrug: